pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcphub"
version = "0.1.8"
description = "A Python package for managing and integrating Model Context Protocol (MCP) servers with AI frameworks like OpenAI Agents, LangChain, and Autogen"
readme = "README.md"
authors = [
    {name = "Hieu TRAN", email = "hieutrantrung.it@gmail.com"},
    {name = "Thang LE", email = "levietthang0512@outlook.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = []
requires-python = "<4.0,>=3.10"

[project.optional-dependencies]
dev = [
    "build>=1.2.2.post1,<2.0.0",
    "twine>=6.1.0,<7.0.0",
]
openai = [
    "openai-agents (>=0.0.9,<0.0.10)",
]
langchain = [
    "langchain-mcp-adapters (>=0.0.7,<0.0.8)",
]
autogen = [
    "autogen-ext[mcp] (>=0.5.1,<0.6.0)",
]
all = [
    "openai-agents (>=0.0.9,<0.0.10)",
    "langchain-mcp-adapters (>=0.0.7,<0.0.8)",
    "autogen-ext[mcp] (>=0.5.1,<0.6.0)",
]

[project.urls]
Documentation = "https://raw.githubusercontent.com/Cognitive-Stack/mcphub/refs/heads/master/README.md"
Issues = "https://github.com/Cognitive-Stack/mcphub/issues"
Source = "https://github.com/Cognitive-Stack/mcphub"

[project.scripts]
mcphub = "mcphub.cli.commands:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"